+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Modifications to existing flavors:
 - Changed RUT to NIT in CONITField form field error message.
 - Fixed validation of Czech birth numbers for birth dates after 1st January 1954
   (`gh-315 <https://github.com/django/django-localflavor/issues/315>`_).
+- Deprecated `nl.validators.NLBankAccountNumberFieldValidator` and `nl.validators.NLPhoneNumberFieldValidator`.
 
 1.6   (2017-11-22)
 ------------------